+++ packages/tabs/src/index.js
@@ -1,3 +1,10 @@
 export { default } from "./Tabs";
 export { default as Tab } from "./Tab";
-export { alignments, AVAILABLE_ALIGNMENTS } from "./alignments";
+export {
+  alignments,
+  variants,
+  orientations,
+  AVAILABLE_ALIGNMENTS,
+  AVAILABLE_VARIANTS,
+  AVAILABLE_ORIENTATIONS
+} from "./constants";