+++ packages/tabs/src/presenters/TabsPresenter.js
@@ -1,17 +1,42 @@
 import React from "react";
-import { css } from "emotion";
+import { cx, css } from "emotion";
 import PropTypes from "prop-types";
 import ThemeContext from "@hig/theme-context";
-import stylesheet from "./TabsPresenter.stylesheet";
-import { alignments, AVAILABLE_ALIGNMENTS } from "../alignments";
+import { createCustomClassNames } from "@hig/utils";
 
-export default function TabsPresenter({ align, children }) {
+import stylesheet from "./TabsPresenter.stylesheet";
+import {
+  AVAILABLE_ALIGNMENTS,
+  AVAILABLE_VARIANTS,
+  AVAILABLE_ORIENTATIONS,
+  alignments,
+  variants,
+  orientations
+} from "../constants";
+
+export default function TabsPresenter({
+  children,
+  align,
+  variant,
+  orientation,
+  ...otherProps
+}) {
   return (
     <ThemeContext.Consumer>
       {({ resolvedRoles }) => {
-        const styles = stylesheet({ align }, resolvedRoles);
+        const styles = stylesheet(
+          { align, variant, orientation },
+          resolvedRoles
+        );
 
-        return <ul className={css(styles.tabs)}>{children}</ul>;
+        const { className } = otherProps;
+        const customClassName = createCustomClassNames(className, "tabs");
+
+        return (
+          <ul className={cx(css(styles.tabsWrapper), customClassName)}>
+            {children}
+          </ul>
+        );
       }}
     </ThemeContext.Consumer>
   );
@@ -19,9 +44,13 @@ export default function TabsPresenter({ align, children }) {
 
 TabsPresenter.propTypes = {
   align: PropTypes.oneOf(AVAILABLE_ALIGNMENTS),
+  variant: PropTypes.oneOf(AVAILABLE_VARIANTS),
+  orientation: PropTypes.oneOf(AVAILABLE_ORIENTATIONS),
   children: PropTypes.node
 };
 
 TabsPresenter.defaultProps = {
-  align: alignments.CENTER
+  align: alignments.LEFT,
+  variant: variants.UNDERLINE,
+  orientation: orientations.HORIZONTAL
 };