+++ packages/theme-data/src/densities/highDensity/components/tabs.js
@@ -1,0 +1,8 @@
+export default {
+  "tabs.box.divider.height": {
+    value: "12px"
+  },
+  "tabs.canvas.divider.height": {
+    value: "12px"
+  }
+};